Reorderd for multiple endspoints, and add async
