--- api/formatter/base_formatter.py
@@ -1,16 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class EDR_formatter(ABC):
-    """
-    This is the abstract class for implementing a formatter in the E-SOH EDR formatter
-    Name of class should represent expected output format.
-    """
-    pass
-
-    @abstractmethod
-    def convert(self, datastore_reply):
-        """
-        Main method for converting protobuf object to given format.
-        """
-        pass
